Delete all display added spotlights, refactor RscDisplayClient to file
